pyproject.toml
[tool.poetry]
name = "git-info"
version = "0.1.0"
description = ""
authors = ["Epsy Health <engineering@epsyhealth.com>"]
packages = [
    { include = "git_info" }
]

[tool.poetry.scripts]
gitinfo = 'git_info.cli:run'

[tool.poetry.dependencies]
python = "^3.8"
click = "^7.1.2"
GitPython = "^3.1.11"
PyGithub = "^1.53"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
